code/backend/requirements.txt
openai
antropic
youtube-transcript-api
flask
flask_cors
python-dotenv
gunicorn
